JANITORIAL: Use size_t and ptrdiff_t more

+++ src/aurora/2dafile.cpp
@@ -50,7 +50,7 @@ TwoDARow::TwoDARow(TwoDAFile &parent) : _parent(&parent) {
 TwoDARow::~TwoDARow() {
 }
 
-const Common::UString &TwoDARow::getString(uint32 column) const {
+const Common::UString &TwoDARow::getString(size_t column) const {
 	const Common::UString &cell = getCell(column);
 	if (cell.empty() || (cell == "****"))
 		return _parent->_defaultString;
@@ -66,7 +66,7 @@ const Common::UString &TwoDARow::getString(const Common::UString &column) const 
 	return cell;
 }
 
-int32 TwoDARow::getInt(uint32 column) const {
+int32 TwoDARow::getInt(size_t column) const {
 	const Common::UString &cell = getCell(column);
 	if (cell.empty() || (cell == "****"))
 		return _parent->_defaultInt;
@@ -82,7 +82,7 @@ int32 TwoDARow::getInt(const Common::UString &column) const {
 	return _parent->parseInt(cell);
 }
 
-float TwoDARow::getFloat(uint32 column) const {
+float TwoDARow::getFloat(size_t column) const {
 	const Common::UString &cell = getCell(column);
 	if (cell.empty() || (cell == "****"))
 		return _parent->_defaultFloat;
@@ -98,7 +98,7 @@ float TwoDARow::getFloat(const Common::UString &column) const {
 	return _parent->parseFloat(cell);
 }
 
-bool TwoDARow::empty(uint32 column) const {
+bool TwoDARow::empty(size_t column) const {
 	const Common::UString &cell = getCell(column);
 	if (cell.empty() || (cell == "****"))
 		return true;
@@ -111,7 +111,7 @@ bool TwoDARow::empty(const Common::UString &column) const {
 }
 
 static const Common::UString kEmpty;
-const Common::UString &TwoDARow::getCell(uint32 n) const {
+const Common::UString &TwoDARow::getCell(size_t n) const {
 	if (n >= _data.size())
 		return kEmpty;
 
@@ -221,14 +221,14 @@ void TwoDAFile::readHeaders2a(Common::SeekableReadStream &twoda,
 void TwoDAFile::readRows2a(Common::SeekableReadStream &twoda,
                            Common::StreamTokenizer &tokenize) {
 
-	uint32 columnCount = _headers.size();
+	size_t columnCount = _headers.size();
 
 	while (!twoda.eos()) {
 		TwoDARow *row = new TwoDARow(*this);
 
 		tokenize.skipToken(twoda);
 
-		int count = tokenize.getTokens(twoda, row->_data, columnCount, columnCount);
+		size_t count = tokenize.getTokens(twoda, row->_data, columnCount, columnCount);
 
 		tokenize.nextChunk(twoda);
 
@@ -272,9 +272,9 @@ void TwoDAFile::skipRowNames2b(Common::SeekableReadStream &twoda) {
 }
 
 void TwoDAFile::readRows2b(Common::SeekableReadStream &twoda) {
-	uint32 columnCount = _headers.size();
-	uint32 rowCount    = _rows.size();
-	uint32 cellCount   = columnCount * rowCount;
+	size_t columnCount = _headers.size();
+	size_t rowCount    = _rows.size();
+	size_t cellCount   = columnCount * rowCount;
 
 	uint32 *offsets = new uint32[cellCount];
 
@@ -282,20 +282,20 @@ void TwoDAFile::readRows2b(Common::SeekableReadStream &twoda) {
 
 	tokenize.addSeparator('\0');
 
-	for (uint32 i = 0; i < cellCount; i++)
+	for (size_t i = 0; i < cellCount; i++)
 		offsets[i] = twoda.readUint16LE();
 
 	twoda.skip(2); // Reserved
 
-	uint32 dataOffset = twoda.pos();
+	size_t dataOffset = twoda.pos();
 
-	for (uint32 i = 0; i < rowCount; i++) {
+	for (size_t i = 0; i < rowCount; i++) {
 		_rows[i] = new TwoDARow(*this);
 
 		_rows[i]->_data.resize(columnCount);
 
-		for (uint32 j = 0; j < columnCount; j++) {
-			uint32 offset = dataOffset + offsets[i * columnCount + j];
+		for (size_t j = 0; j < columnCount; j++) {
+			size_t offset = dataOffset + offsets[i * columnCount + j];
 
 			try {
 				twoda.seek(offset);
@@ -314,15 +314,15 @@ void TwoDAFile::readRows2b(Common::SeekableReadStream &twoda) {
 }
 
 void TwoDAFile::createHeaderMap() {
-	for (uint32 i = 0; i < _headers.size(); i++)
+	for (size_t i = 0; i < _headers.size(); i++)
 		_headerMap.insert(std::make_pair(_headers[i], i));
 }
 
-uint32 TwoDAFile::getRowCount() const {
+size_t TwoDAFile::getRowCount() const {
 	return _rows.size();
 }
 
-uint32 TwoDAFile::getColumnCount() const {
+size_t TwoDAFile::getColumnCount() const {
 	return _headers.size();
 }
 
@@ -330,7 +330,7 @@ const std::vector<Common::UString> &TwoDAFile::getHeaders() const {
 	return _headers;
 }
 
-uint32 TwoDAFile::headerToColumn(const Common::UString &header) const {
+size_t TwoDAFile::headerToColumn(const Common::UString &header) const {
 	HeaderMap::const_iterator column = _headerMap.find(header);
 	if (column == _headerMap.end())
 		// No such header
@@ -339,7 +339,7 @@ uint32 TwoDAFile::headerToColumn(const Common::UString &header) const {
 	return column->second;
 }
 
-const TwoDARow &TwoDAFile::getRow(uint32 row) const {
+const TwoDARow &TwoDAFile::getRow(size_t row) const {
 	if ((row >= _rows.size()) || !_rows[row])
 		// No such row
 		return _emptyRow;
@@ -361,21 +361,21 @@ bool TwoDAFile::dumpASCII(const Common::UString &fileName) const {
 
 	// Calculate column lengths
 
-	std::vector<uint32> colLength;
+	std::vector<size_t> colLength;
 	colLength.resize(_headers.size() + 1, 0);
 
 	const Common::UString maxRow = Common::UString::format("%d", (int)_rows.size() - 1);
 	colLength[0] = maxRow.size();
 
-	for (uint32 i = 0; i < _headers.size(); i++)
+	for (size_t i = 0; i < _headers.size(); i++)
 		colLength[i + 1] = _headers[i].size();
 
-	for (uint32 i = 0; i < _rows.size(); i++) {
-		for (uint32 j = 0; j < _rows[i]->_data.size(); j++) {
+	for (size_t i = 0; i < _rows.size(); i++) {
+		for (size_t j = 0; j < _rows[i]->_data.size(); j++) {
 			const bool   needQuote = _rows[i]->_data[j].contains(' ');
-			const uint32 length    = needQuote ? _rows[i]->_data[j].size() + 2 : _rows[i]->_data[j].size();
+			const size_t length    = needQuote ? _rows[i]->_data[j].size() + 2 : _rows[i]->_data[j].size();
 
-			colLength[j + 1] = MAX<uint32>(colLength[j + 1], length);
+			colLength[j + 1] = MAX<size_t>(colLength[j + 1], length);
 		}
 	}
 
@@ -383,17 +383,17 @@ bool TwoDAFile::dumpASCII(const Common::UString &fileName) const {
 
 	file.writeString(Common::UString::format("%-*s", colLength[0], ""));
 
-	for (uint32 i = 0; i < _headers.size(); i++)
+	for (size_t i = 0; i < _headers.size(); i++)
 		file.writeString(Common::UString::format(" %-*s", colLength[i + 1], _headers[i].c_str()));
 
 	file.writeByte('\n');
 
 	// Write array
 
-	for (uint32 i = 0; i < _rows.size(); i++) {
+	for (size_t i = 0; i < _rows.size(); i++) {
 		file.writeString(Common::UString::format("%*d", colLength[0], i));
 
-		for (uint32 j = 0; j < _rows[i]->_data.size(); j++) {
+		for (size_t j = 0; j < _rows[i]->_data.size(); j++) {
 			const bool needQuote = _rows[i]->_data[j].contains(' ');
 
 			Common::UString cellString;
